+++ src/main/java/ch/epfl/cs107/play/game/icrogue/actor/enemies/Turret.java
@@ -21,7 +21,7 @@ public class Turret extends Enemy{
     private boolean shootRight;
     public final static float COOLDOWN = 2.f;
     private float counter = 0;
-    private float hp = 5;
+    private int hp = 5;
 
     public Turret(Area owner, Orientation orientation, DiscreteCoordinates coordinates, boolean sUp
             ,boolean sDown, boolean sLeft, boolean sRight) {
@@ -90,7 +90,7 @@ public class Turret extends Enemy{
         super.update(deltaTime);
     }
 
-    public void decreaseHp(float delta){
+    public void decreaseHp(int delta){
         hp -= delta;
         System.out.println("turret: "+hp);
     }
